Extract AllReferencesToAFileTestHelper
