+++ java/core/src/test/java/sleeper/core/statestore/AllReferencesToAFileTestHelper.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2022-2024 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sleeper.core.statestore;
+
+import java.time.Instant;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+public class AllReferencesToAFileTestHelper {
+
+    private AllReferencesToAFileTestHelper() {
+    }
+
+    public static AllReferencesToAFile fileWithNoReferences(String filename) {
+        return fileWithNoReferences(filename, null);
+    }
+
+    public static AllReferencesToAFile fileWithNoReferences(String filename, Instant updateTime) {
+        return AllReferencesToAFile.builder()
+                .filename(filename)
+                .internalReferenceByPartitionId(Map.of())
+                .totalReferenceCount(0)
+                .lastStateStoreUpdateTime(updateTime)
+                .build();
+    }
+
+    public static AllReferencesToAFile fileWithReferences(Collection<FileReference> references) {
+        List<AllReferencesToAFile> files = listNewFilesWithReferences(references, null);
+        if (files.size() != 1) {
+            throw new IllegalArgumentException("Expected one file, found: " + files);
+        }
+        return files.get(0);
+    }
+
+    public static List<AllReferencesToAFile> listNewFilesWithReferences(Collection<FileReference> references, Instant updateTime) {
+        return newFilesWithReferences(references, updateTime)
+                .collect(Collectors.toUnmodifiableList());
+    }
+
+    public static Stream<AllReferencesToAFile> newFilesWithReferences(Collection<FileReference> references, Instant updateTime) {
+        return AllReferencesToAFile.newFilesWithReferences(references.stream(), updateTime);
+    }
+
+}
