+++ java/core/src/test/java/sleeper/core/statestore/FilesReportTestHelper.java
@@ -35,7 +35,7 @@ public class FilesReportTestHelper {
     }
 
     public static AllReferencesToAllFiles activeFilesReport(Instant updateTime, List<FileReference> references) {
-        return new AllReferencesToAllFiles(AllReferencesToAFile.listNewFilesWithReferences(references, updateTime), false);
+        return new AllReferencesToAllFiles(AllReferencesToAFileTestHelper.listNewFilesWithReferences(references, updateTime), false);
     }
 
     public static AllReferencesToAllFiles activeAndReadyForGCFilesReport(
@@ -54,14 +54,9 @@ public class FilesReportTestHelper {
     private static List<AllReferencesToAFile> activeAndReadyForGCFiles(
             Instant updateTime, List<FileReference> activeFiles, List<String> readyForGCFiles) {
         return Stream.concat(
-                AllReferencesToAFile.newFilesWithReferences(activeFiles, updateTime),
+                AllReferencesToAFileTestHelper.newFilesWithReferences(activeFiles, updateTime),
                 readyForGCFiles.stream().map(filename ->
-                        AllReferencesToAFile.builder()
-                                .filename(filename)
-                                .internalReferences(List.of())
-                                .totalReferenceCount(0)
-                                .lastStateStoreUpdateTime(updateTime)
-                                .build())
+                        AllReferencesToAFileTestHelper.fileWithNoReferences(filename, updateTime))
         ).collect(Collectors.toUnmodifiableList());
     }
 }
